kubevirt: move vm actions request
